[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -8,7 +8,7 @@ int main() {
     ios::sync_with_stdio(0);
     cin.tie(0);
     
-    int x, y; cin >> x >> y;
+    ll x, y; cin >> x >> y;
     
     cout << (x + y) * (x - y);
